moving to correct package structure for loading plugins
