+++ data-prepper-plugins/failures-common/src/main/java/org/opensearch/dataprepper/plugins/dlq/DlqWriter.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.dataprepper.dlq;
+package org.opensearch.dataprepper.plugins.dlq;
 
 
 import org.opensearch.dataprepper.model.failures.DlqObject;
